[GREEN] Move dashboard download logging to curator feature
