+++ curator_feature/serializers.py
@@ -1,6 +1,13 @@
 from rest_framework import serializers
 
-from curator_feature.models import DownloadLog
+from curator_feature.models import DownloadLog, DashboardDownloadEvent
+
+
+class CaseInsensitiveChoiceField(serializers.ChoiceField):
+    def to_internal_value(self, data):
+        if isinstance(data, str):
+            data = data.lower()
+        return super().to_internal_value(data)
 
 
 class DownloadLogRequestSerializer(serializers.Serializer):
@@ -25,3 +32,10 @@ class DownloadLogResponseSerializer(serializers.ModelSerializer):
     class Meta:
         model = DownloadLog
         fields = ("id", "username", "chartType", "timestamp", "created_at")
+
+
+class DashboardDownloadEventSerializer(serializers.Serializer):
+    metric = CaseInsensitiveChoiceField(choices=DashboardDownloadEvent.Metric.choices)
+    file_format = CaseInsensitiveChoiceField(choices=DashboardDownloadEvent.FileFormat.choices)
+    filters = serializers.JSONField(required=False)
+    source = serializers.CharField(required=False, allow_blank=True)
