+++ pt_backend/serializers.py
@@ -1,13 +1,5 @@
 from rest_framework import serializers
 from .constants import PROVINCE_TO_CODE
-from .models import DownloadEvent
-
-
-class CaseInsensitiveChoiceField(serializers.ChoiceField):
-    def to_internal_value(self, data):
-        if isinstance(data, str):
-            data = data.lower()
-        return super().to_internal_value(data)
 
 class CaseLocationSerializer(serializers.Serializer):
     id = serializers.UUIDField()
@@ -72,13 +64,3 @@ class ProvinceClimateSerializer(serializers.Serializer):
 ProvinceHumiditySerializer = ProvinceClimateSerializer
 ProvinceTemperatureSerializer = ProvinceClimateSerializer
 ProvincePrecipitationSerializer = ProvinceClimateSerializer
-
-
-class DownloadLogSerializer(serializers.Serializer):
-    metric = CaseInsensitiveChoiceField(choices=DownloadEvent.Metric.choices)
-    file_format = CaseInsensitiveChoiceField(choices=DownloadEvent.FileFormat.choices)
-    filters = serializers.JSONField(required=False)
-    source = serializers.CharField(required=False, allow_blank=True)
-
-    def validate_file_format(self, value):
-        return value.lower()
